rec engine fixes

+++ pet-adoption-api/src/main/java/petadoption/api/adoptioncenter/AdoptionCenterService.java
@@ -27,7 +27,7 @@ public class AdoptionCenterService {
         return new ResponseEntity<>(centers, HttpStatus.OK);
     }
 
-    public ResponseEntity<AdoptionCenter> getAdoptionCenter(int id)  {
+    public ResponseEntity<AdoptionCenter> getAdoptionCenter(Long id)  {
         Optional<AdoptionCenter> center = adoptionCenterRepository.findById((long) id);
         return center
                 .map(value -> new ResponseEntity<>(value, HttpStatus.OK))
